core/api_datapii/requirements.txt
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.1.0
Requests==2.32.3
